backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
Pillow>=10.0.0
pytesseract>=0.3.10
requests>=2.31.0
pydantic>=2.0.0
